requirements.txt
authlib
dynaconf
globus-compute-sdk @ git+https://github.com/globus/globus-compute.git@d1731340074be56861ec91d732bdff44f8e2b46e#subdirectory=compute_sdk
globus-sdk>=3.0
griffe>=0.49.0,<2.0.0
h5py
httpx>=0.22.0
mkdocs
mkdocs-material
mkdocs-mermaid2-plugin
numpy>=1.26.4
pillow
pydantic==2.11
python-dotenv
prefect==2.20.17
prometheus_client==0.21.1
pyscicat
pyyaml
sfapi_client